cleaned up the tests
